chess_engine/requirements.txt
Django==5.0.6
gunicorn==22.0.0
whitenoise==6.7.0
opencv-python-headless==4.10.0.84
python-chess==1.999
board_to_fen
tensorflow-cpu
Pillow